refactor(ipc): consolidate shared types and eliminate duplication
Create shared type definitions to follow DRY principle:

- session/types.ts (new) - Shared types: SessionActivity, PageState
- session/lifecycle.ts - Add SessionOptions type
- protocol/commands.ts - Use shared types instead of inline definitions
- client.ts - Use SessionOptions instead of inline parameter types

Benefits:
- Single source of truth for common types
- Easier to maintain and update type definitions
- Better type reusability across modules
- Reduces code duplication (22 lines removed)
