+++ src/ipc/session/queries.ts
@@ -5,32 +5,9 @@
  */
 
 import type { IPCMessage } from './lifecycle.js';
+import type { PageState, SessionActivity } from './types.js';
 
 import type { TelemetryType } from '@/types.js';
-
-/**
- * Session activity metrics.
- */
-export interface SessionActivity {
-  /** Total network requests captured. */
-  networkRequestsCaptured: number;
-  /** Total console messages captured. */
-  consoleMessagesCaptured: number;
-  /** Timestamp of last network request. */
-  lastNetworkRequestAt?: number;
-  /** Timestamp of last console message. */
-  lastConsoleMessageAt?: number;
-}
-
-/**
- * Current page state.
- */
-export interface PageState {
-  /** Current page URL. */
-  url: string;
-  /** Current page title. */
-  title: string;
-}
 
 /**
  * Status request (client → daemon).
